refactor(rag): extract review package
